pyproject.toml
[tool.poetry]
authors = ["HB6091 <HB6091@engie.com>"]
description = "test"
name = "cdh_test"
version = "0.5.0"

[tool.poetry.dependencies]
fastapi = "^0.75.1"
gunicorn = "^20.1.0"
python = ">=3.9.0,<3.11"
# https://github.com/tiangolo/fastapi/issues/2181
azure-keyvault-secrets = "^4.2.0"
pandas = "^1.2.5"
pytest = "^6.2.3"
requests-auth = "^5.3.0"
requests_ntlm = "^1.1.0"
toml = "^0.10.2"
python-certifi-win32 = "^1.6.1"
azure-keyvault = "^4.2.0"
azure-identity = "^1.10.0"
PyYAML = "^6.0"
tabulate = "^0.8.10"
boto3 = "^1.24.70"
awswrangler = "^2.16.1"

[tool.poetry.dev-dependencies]
black = {version = "^20.8b1", allow-prereleases = true}

[tool.black]
line-length = 119

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
